+++ skia-safe/src/core/milestone.rs
@@ -1,1 +1,1 @@
-pub const MILESTONE: usize = 126;
+pub const MILESTONE: usize = 127;
